Moved ServiceWatcherUtil into asynchronous.management
